refactor(mcp) - create example server directory
