+++ src/example_weather_mcp/constants.py
